# api/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OctoCAT Supply Chain Management API",
        "version": "1.0.0",
        "description": "A comprehensive REST API for managing supply chain operations including suppliers, products, orders, deliveries, and branch locations. Built with Express.js, TypeScript, and SQLite.",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/webmaxru/copilot_agent_mode"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        },
        {
            "url": "https://localhost:3000",
            "description": "Development server (HTTPS)"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "summary": "Returns all products",
                "tags": [
                    "Products"
                ],
                "responses": {
                    "200": {
                        "description": "List of all products",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new product",
                "tags": [
                    "Products"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "summary": "Get a product by ID",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "summary": "Update a product",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Product"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a product",
                "tags": [
                    "Products"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Product ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/suppliers": {
            "get": {
                "summary": "Returns all suppliers",
                "tags": [
                    "Suppliers"
                ],
                "responses": {
                    "200": {
                        "description": "List of all suppliers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Supplier"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new supplier",
                "tags": [
                    "Suppliers"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Supplier"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplier created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/suppliers/{id}": {
            "get": {
                "summary": "Get a supplier by ID",
                "tags": [
                    "Suppliers"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Supplier ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplier found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            },
            "put": {
                "summary": "Update a supplier",
                "tags": [
                    "Suppliers"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Supplier ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Supplier"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplier updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Supplier"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete a supplier",
                "tags": [
                    "Suppliers"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Supplier ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supplier deleted successfully"
                    },
                    "404": {
                        "description": "Supplier not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Product": {
                "type": "object",
                "required": [
                    "productId",
                    "name",
                    "price"
                ],
                "properties": {
                    "productId": {
                        "type": "integer",
                        "description": "The unique identifier for the product"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the product"
                    },
                    "description": {
                        "type": "string",
                        "description": "Detailed description of the product"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "description": "The current price of the product"
                    },
                    "supplierId": {
                        "type": "integer",
                        "description": "The ID of the supplier providing this product"
                    },
                    "stockLevel": {
                        "type": "integer",
                        "description": "Current stock level of the product"
                    }
                }
            },
            "Supplier": {
                "type": "object",
                "required": [
                    "supplierId",
                    "name"
                ],
                "properties": {
                    "supplierId": {
                        "type": "integer",
                        "description": "The unique identifier for the supplier"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the supplier"
                    },
                    "address": {
                        "type": "string",
                        "description": "The physical address of the supplier"
                    },
                    "contactPerson": {
                        "type": "string",
                        "description": "Name of the primary contact person"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "description": "Contact email for the supplier"
                    },
                    "phone": {
                        "type": "string",
                        "description": "Contact phone number for the supplier"
                    },
                    "description": {
                        "type": "string",
                        "description": "Additional details about the supplier"
                    }
                }
            },
...
        }
    },
...
}